refactor(controller): rename package `controller` -> `controllers`
